Challenge Route, rework of interface

+++ src/routes/challenge/controller.ts
@@ -7,13 +7,18 @@ import {
 
 export async function createChallenge(
   fastify: FastifyInstance,
-  challengeInfo: ChallengeInfo
+  challengeInfo: ChallengeInfo,
+  creatorId: number
 ) {
   if (challengeInfo.reward && challengeInfo.reward < 0) {
     throw fastify.httpErrors.badRequest("Reward must pe positive");
   }
 
-  await fastify.prisma.challenge.createChallenge(challengeInfo);
+  if (!creatorId) {
+    throw fastify.httpErrors.badRequest("Invalid creator id");
+  }
+
+  await fastify.prisma.challenge.createChallenge(challengeInfo, creatorId);
 }
 
 export async function updateChallenge(
@@ -25,6 +30,10 @@ export async function updateChallenge(
     throw fastify.httpErrors.badRequest("Reward must pe positive");
   }
 
+  if (!challengeId) {
+    throw fastify.httpErrors.badRequest("Invalid challenge id");
+  }
+
   await fastify.prisma.challenge.updateChallenge(challengeInfo, challengeId);
 }
 
@@ -32,6 +41,10 @@ export async function deleteChallenge(
   fastify: FastifyInstance,
   challengeId: number
 ) {
+  if (!challengeId) {
+    throw fastify.httpErrors.badRequest("Invalid challenge id");
+  }
+
   await fastify.prisma.challenge.deleteChallenge(challengeId);
 }
 
@@ -39,6 +52,10 @@ export async function getChallenge(
   fastify: FastifyInstance,
   challengeId: number
 ) {
+  if (!challengeId) {
+    throw fastify.httpErrors.badRequest("Invalid challenge id");
+  }
+
   const challenge = await fastify.prisma.challenge.getChallenge(challengeId);
 
   //Check if challenge not found
